move capture_images
